requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.6.1
python-dotenv==1.0.0
asyncio-mqtt==0.13.0
aiofiles==23.2.1
pillow==10.2.0
qrcode==7.4.2
jdatetime==4.1.0
redis==5.0.1
bcrypt==4.1.2
cryptography==41.0.8
schedule==1.2.0
psutil==5.9.6
aiohttp==3.9.1
motor==3.3.2